Fix Dayline v0.12.2 timeline build

+++ app/src/main/java/com/pix/dayline/ui/components/DayTimeline.kt
@@ -409,7 +409,7 @@ private fun TimelineGap(
     onCreateAt: (LocalDate, LocalTime) -> Unit
 ) {
     val total = Duration.between(from, to).toMinutes().coerceAtLeast(15L)
-    val height = (total / 15L * 5L).coerceIn(16L, 52L).dp
+    val height = (total / 15L * 5L).coerceIn(16L, 52L).toInt().dp
 
     Box(
         modifier = Modifier
